Projet Green Form/greenform_project/requirement.txt
django==3.2.6
mysqlclient==2.0.3
pandas
pillow
qrcode